+++ src/main/java/us/hogu/controller/dto/request/BnbPriceCalendarRequestDto.java
@@ -1,5 +1,6 @@
 package us.hogu.controller.dto.request;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import lombok.Data;
@@ -8,5 +9,5 @@ import lombok.Data;
 public class BnbPriceCalendarRequestDto {
     private LocalDate startDate;
     private LocalDate endDate;
-    private Double pricePerNight;
+    private BigDecimal pricePerNight;
 }
