+++ src/main/java/us/hogu/controller/dto/request/BnbRoomRequestDto.java
@@ -1,5 +1,6 @@
 package us.hogu.controller.dto.request;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import lombok.Data;
@@ -9,7 +10,7 @@ public class BnbRoomRequestDto {
     private String name;
     private String description;
     private Integer maxGuests;
-    private Double basePricePerNight;
+    private BigDecimal basePricePerNight;
     private Boolean available;
     private List<String> images;
 }
