+++ src/main/java/us/hogu/controller/dto/request/BnbServiceRequestDto.java
@@ -1,5 +1,6 @@
 package us.hogu.controller.dto.request;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import lombok.Data;
@@ -8,7 +9,7 @@ import lombok.Data;
 public class BnbServiceRequestDto {
     private String name;
     private String description;
-    private Double defaultPricePerNight;
+    private BigDecimal defaultPricePerNight;
     private Integer totalRooms;
     private Integer maxGuestsForRoom;
     private List<ServiceLocaleRequestDto> locales;
