+++ src/main/java/us/hogu/controller/dto/request/LuggageSizePriceRequestDto.java
@@ -1,5 +1,7 @@
 package us.hogu.controller.dto.request;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -11,8 +13,8 @@ import lombok.NoArgsConstructor;
 @Builder
 public class LuggageSizePriceRequestDto {
     private String sizeLabel;
-    private Double pricePerHour;
-    private Double pricePerDay;
+    private BigDecimal pricePerHour;
+    private BigDecimal pricePerDay;
     private String description;
 }
