+++ src/main/java/us/hogu/controller/dto/response/LuggageSizePriceResponseDto.java
@@ -1,5 +1,7 @@
 package us.hogu.controller.dto.response;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -12,8 +14,8 @@ import lombok.NoArgsConstructor;
 public class LuggageSizePriceResponseDto {
     private Long id;
     private String sizeLabel;       // SMALL, MEDIUM, LARGE, EXTRA_LARGE
-    private Double pricePerHour;
-    private Double pricePerDay;
+    private BigDecimal pricePerHour;
+    private BigDecimal pricePerDay;
     private String description;     // es. "fino a 15 kg"
 }
